backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
python-dotenv==1.0.0
gunicorn==21.2.0venv\Scripts\activate
